# swagger.yaml
swagger: '2.0'
info:
  description: How to use the Boastable server.
  version: 1.0.0
  title: Boastable API
  contact:
    email: ryan@fueledonbacon.com
host: localhost:3000
basePath: /
...
schemes:
  - http
paths:
...
  /auth:
    post:
      tags:
        - auth
      summary: Create auth document
      description: Clients create their own customers
      operationId: createAuth
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          description: an authorization header
          example: bearer [insert JWT token here]
          required: true
          type: string
        - in: body
          name: body
          description: Created auth document
          required: true
          schema:
            $ref: '#/definitions/Auth'
      responses:
        '200':
          description: OK
    get:
      tags:
        - auth
      summary: Get customer
      description: ''
      operationId: getAuth
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          description: an authorization header
          example: bearer [insert JWT token here]
          required: true
          type: string
        - name: id
          in: query
          description: get auth document for a client
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Auth'
        '400':
          description: Invalid username supplied
        '404':
          description: Auth not found
    put:
      tags:
        - auth
      summary: Update customer data
      description: This can only be done by the logged in user.
      operationId: updateAuth
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          description: an authorization header
          example: bearer [insert JWT token here]
          required: true
          type: string
        - name: id
          in: query
          description: ID of the customer that needs to be updated
          required: true
          type: string
        - in: body
          name: body
          description: Updated customer object
          required: true
          schema:
            $ref: '#/definitions/Auth'
      responses:
        '200':
          description: OK
        '400':
          description: Invalid user supplied
        '404':
          description: Auth not found
    delete:
      tags:
        - auth
      summary: Delete auth document
      description: Only an admin can do this
      operationId: deleteAuth
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          description: an authorization header
          example: bearer [insert JWT token here]
          required: true
          type: string
        - name: id
          in: query
          description: A customer that needs to be deleted
          required: true
          type: string
      responses:
        '200':
          description: OK
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '404':
          description: Auth not found
...
definitions:
...
  Auth:
    type: object
    required:
      - email
      - password
    properties:
      _id:
        type: string
        description: ObjectId
        example: 5a2cfe672ebedf740dac80c8
      email:
        type: string
        example: example@user.com
      password:
        type: string
        description: hash of plaintext password
...